+++ ScavgameTranslationUtils/App.axaml.cs
@@ -1,7 +1,6 @@
+using System.IO;
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
-using Avalonia.Data.Core.Plugins;
-using System.Linq;
 using Avalonia.Markup.Xaml;
 using ScavgameTranslationUtils.ViewModels;
 using ScavgameTranslationUtils.Views;
@@ -19,10 +18,22 @@ public partial class App : Application
     {
         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
         {
+            MainWindowViewModel viewModel;
+            
             desktop.MainWindow = new MainWindow
             {
-                DataContext = new MainWindowViewModel(),
+                DataContext = viewModel = new MainWindowViewModel(),
             };
+            
+            // TODO: Hacky?
+            Avalonia.Threading.Dispatcher.UIThread.InvokeAsync(async () =>
+            {
+                if (File.Exists(Program.Settings.GameExecutablePath))
+                    await viewModel.OpenGameAssets(Program.Settings.GameExecutablePath);
+        
+                if (File.Exists(Program.Settings.EnglishTranslationPath))
+                    await viewModel.OpenOriginalTranslation(Program.Settings.EnglishTranslationPath);
+            });
         }
 
         base.OnFrameworkInitializationCompleted();